requirements.txt
fastapi
gunicorn
uvicorn
langchain
duckduckgo-search
python-dotenv
google-generativeai
langchain-community
langchain-google-genai



